+++ changelog.md
@@ -1,7 +1,16 @@
 
 # Changelog
 
-DataSciproLab http://datasciprolab.readthedocs.com
+Scientific Programming Data Science Lab 
+
+https://sciprog.davidleoni.it
+
+
+## 1.0 September 2020
+
+- upgraded to Jupman 3, changed folder structure
+- moved webste to sciprog.davidleoni.it
+- building with Github Actions
 
 
 ## 0.1, September 2018